-Fixing bugs to create the dot files
-Creating the dot files on a different folder (they were previous being created on cfg folder)
-Adding some prints on every loop, so I know how much of it ran
-Adding the anghaMath programs in the compile stage

+++ histogramGenerator.py
@@ -1,9 +1,11 @@
 import os
+import subprocess
 
 def main():
 	compileFilesFrom("./jotai-benchmarks/benchmarks/anghaLeaves", "./binaries")
+	compileFilesFrom("./jotai-benchmarks/benchmarks/anghaMath", "./binaries")
 	generateCFGgrindMapsFrom("./binaries", "./maps")
-	generateDotFilesFrom("./binaries", "./cfg")
+	generateDotFilesFrom("./binaries", "./cfg", "./maps", "./dots")
 
 def compileFilesFrom(folderPathPrograms, folderPathBinaries):
 	files = getFilesFrom(folderPathPrograms)
@@ -13,7 +15,7 @@ def compileFilesFrom(folderPathPrograms, folderPathBinaries):
 	for fileName in files:
 		count = count + 1
 		if fileName not in filesNotCompilable():
-			print("Compilando arquivo " + fileName + " -> " + str(count) + " / " + str(len(files)))
+			print("Compiling file " + fileName + " -> " + str(count) + " / " + str(len(files)))
 			regularCompilation(fileName, folderPathPrograms, folderPathBinaries)
 
 def filesNotCompilable():
@@ -26,7 +28,6 @@ def createFolderIfNotExist(folderPath):
 
 def regularCompilation(fileName, folderPathPrograms, folderPathBinaries):
 	osResult = os.system("clang -Wno-everything " + folderPathPrograms + "/" + fileName)
-	print("Result: " + str(osResult) + " | " + fileName + " to " + fileName[:-2] + "-regular.out")
 	if osResult == 0:
 		os.system("cp " + "./a.out " + folderPathBinaries + "/" + fileName[:-2] + "-regular.out")
 
@@ -36,27 +37,38 @@ def getFilesFrom(folderPath):
 def generateCFGgrindMapsFrom(folderPathBinaries, folderPathMaps):
 	files = getFilesFrom(folderPathBinaries)
 	createFolderIfNotExist(folderPathMaps)
+	count = 0
 	for fileName in files:
+		count = count + 1
+		print("Generating map file " + fileName + " -> " + str(count) + " / " + str(len(files)))
 		cfggrindAsMap(fileName, folderPathBinaries, folderPathMaps)
 
 def cfggrindAsMap(fileName, folderPathBinaries, folderPathMaps):
+	print("cfggrind_asmmap " + folderPathBinaries + "/" + fileName + " > " + folderPathMaps + "/" + fileName[:-4] + ".map")
 	os.system("cfggrind_asmmap " + folderPathBinaries + "/" + fileName + " > " + folderPathMaps + "/" + fileName[:-4] + ".map")
 
-def generateDotFilesFrom(folderPathBinaries, folderCfgFiles):
+def generateDotFilesFrom(folderPathBinaries, folderCfgFiles, folderPathMaps, folderPathDots):
 	files = getFilesFrom(folderPathBinaries)
 	createFolderIfNotExist(folderCfgFiles)
+	createFolderIfNotExist(folderPathDots)
+	count = 0
 	for fileName in files:
 		try:
-			generateDotFile(fileName, folderPathBinaries, folderCfgFiles)
+			count = count + 1
+			print("Generating dot file " + fileName + " -> " + str(count) + " / " + str(len(files)))
+			generateDotFile(fileName, folderPathBinaries, folderCfgFiles, folderPathMaps, folderPathDots)
 		except Exception as e: 
 			print(e)
 
-def generateDotFile(fileName, folderPathBinaries, folderPathCfgs):
-	os.system(" valgrind --tool=cfggrind --cfg-outfile=test.cfg --instrs-map=test.map --cfg-dump=bubble " + folderPathBinaries + "/" + fileName + " 0")
-	direct_output = subprocess.check_output('ls *.dot', shell=True)
+def generateDotFile(fileName, folderPathBinaries, folderPathCfgs, folderPathMaps, folderPathDots):
+	fileCfgName = folderPathCfgs + "/" + fileName[:-4] + ".cfg"
+	fileMapName = folderPathMaps + "/" + fileName[:-4] + ".map"
+	os.system(" valgrind --tool=cfggrind --cfg-outfile=" + fileCfgName + " --instrs-map=" + fileMapName + " --cfg-dump=main " + folderPathBinaries + "/" + fileName + " 0")
+	direct_output = subprocess.check_output('ls *.dot', shell=True).decode("utf-8").split("\n")
 	for dotfile in direct_output:
-		os.system("cp " + "./" + dotfile + " " + folderPathCfgs + "/" + fileName[:-4] + "-regular.out")
-		os.system("rm ./" + dotfile)
+		if dotfile != "":
+			os.system("cp " + "./" + dotfile + " " + folderPathDots + "/" + fileName[:-4] + ".dot")
+			os.system("rm ./" + dotfile)
 
 
 if __name__ == "__main__":
